Move CsvGenerator into concern for future refactoring
