+++ dpc-web/app/controllers/internal/users_controller.rb
@@ -4,6 +4,7 @@ require 'csv'
 
 module Internal
   class UsersController < ApplicationController
+    include CsvGenerator
     before_action :authenticate_internal_user!
 
     def index
@@ -57,30 +58,6 @@ module Internal
     end
 
     private
-
-
-    ATTRS = %w[id first_name last_name email requested_organization requested_organization_type
-      address_1 address_2 city state zip agree_to_terms requested_num_providers created_at updated_at].freeze
-
-    # html escape these fields for XSS protection
-    ESCAPED_ATTRS = %w[first_name last_name requested_organization address_1 address_2 city].freeze
-
-    def cvs_convert(users)
-      CSV.generate(headers:true) do |csv|
-        csv << ATTRS
-        users.each do |user|
-          attributes = user.attributes
-          escaped_attributes = attributes.map do |k, v|
-            if ESCAPED_ATTRS.include? k
-              v = ERB::Util.html_escape(v)
-            end
-  
-            [k, v]
-          end.to_h
-          csv << escaped_attributes.values_at(*ATTRS)
-        end
-      end
-    end
 
     def user_params
       params.require(:user).permit(:first_name, :last_name, :email, :organization_ids)
